Update RabinKarp hashing algorithm
Updated the RabinKarp.cpp to optimize the string hashing method. Implemented a new method for calculating the hash of the next substring in the text. The change reduces the time complexity by avoiding needless rehashing for each substring. Also, removed a debug line which printed the original text and pattern.